requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3